working on scope/inspection v2
